+++ src/common/get_README.ts
@@ -2,7 +2,7 @@
 // research later https://github.com/matiassingers/awesome-readme
 // "Give it a star!" ?
 
-import type { Core } from '../core';
+import type { Core } from '../core/core';
 
 
 type Badges = {
@@ -11,12 +11,12 @@ type Badges = {
   typescript?: boolean;
   /** npm package version and downloads/week */
   npm?: boolean;
-}
+};
 
 export type get_README_Options = {
   /** All badges default to false. */
-  badges?: Badges
-}
+  badges?: Badges;
+};
 
 // Be sure when changing the defaults.
 export function get_README(core: Core, options?: get_README_Options): string {
@@ -26,16 +26,18 @@ export function get_README(core: Core, options?: get_README_Options): string {
   function getBadgesString() {
     const badges = options?.badges;
     const badgesStrings = [];
-    if (badges?.prWelcome) badgesStrings.push(
-      `[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)`,
+    if (badges?.npm) badgesStrings.push(
+      `[![npm](https://img.shields.io/npm/v/${projectName})](https://www.npmjs.com/package/${projectName})`,
     );
     if (badges?.typescript) badgesStrings.push(
       `[![TypeScript](https://badgen.net/npm/types/env-var)](http://www.typescriptlang.org/)`,
     );
-    if (badges?.npm) badgesStrings.push(...[
-      `[![npm](https://img.shields.io/npm/v/${projectName})](https://www.npmjs.com/package/${projectName})`,
+    if (badges?.prWelcome) badgesStrings.push(
+      `[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)`,
+    );
+    if (badges?.npm) badgesStrings.push(
       `[![npm](https://img.shields.io/npm/dw/${projectName})](https://www.npmjs.com/package/${projectName})`,
-    ]);
+    );
 
     if (!badgesStrings.length)
       return '';
@@ -43,36 +45,35 @@ export function get_README(core: Core, options?: get_README_Options): string {
     return (`<div align="center">\n\n${badgesStrings.join('\n')}\n</div>`);
   }
 
-
-  return (
-    `
-<h1 align="center">
+  // Is trimmed on end.
+  let result = (`
   <!-- <img src=".logo.png" alt=${projectName}/><br/> -->
+
+  ${getBadgesString()}
+
   ${projectName}
-</h1>
 
-${getBadgesString()}
+  <!-- descriptionHere -->
 
-<br/>
+  <br/>
 
-<div align="center">
-  <h3> 🏗 ❗ Work In Progress ❗🛠 </h3>
-</div>
+  <div align="center">
+    <h3> 🏗 ❗ Work In Progress ❗🛠 </h3>
+  </div>
 
-<br/>
+  <br/>
 
-# Installation
-\`\`\`bash
-npm install ${core.consts.projectName}
-# or
-yarn add ${core.consts.projectName}
-\`\`\`
+  ## Installation
+  \`\`\`bash
+  npm install ${core.consts.projectName}
+  # or
+  yarn add ${core.consts.projectName}
+  \`\`\`
 
-# Usage
+  ## Usage
 
-# [Changelog](CHANGELOG.md)`
-  );
+  ## [Changelog](CHANGELOG.md)`);
+
+  result = result.trim();
+  return result;
 }
-
-// [![npm](https://img.shields.io/npm/v/react-native-shadow-2)](https://www.npmjs.com/package/react-native-shadow-2)
-// [![npm](https://img.shields.io/npm/dt/react-native-shadow-2)](https://www.npmjs.com/package/react-native-shadow-2)