+++ src/flavors/ts.ts
@@ -2,17 +2,36 @@ import type { FlavorFunction } from '../typesAndConsts';
 import editJsonFile from 'edit-json-file';
 import ora from 'ora';
 
-export const flavorTypescript: FlavorFunction = async (core) => {
+
+
+export const typescriptDevDeps = [
+  'typescript@latest',
+  'ts-node-dev@latest',
+
+  'eslint@latest',
+  'eslint-config-gev@latest',
+  '@typescript-eslint/eslint-plugin@latest',
+  '@typescript-eslint/parser@latest',
+
+  '@types/node',
+
+  'jest',
+  'ts-jest',
+  '@types/jest',
+];
+
+
+const flavorTypescript: FlavorFunction = async (core) => {
 
   core.verifications.projectNameMustBeNpmValid();
 
   await core.verifications.projectPathMustBeValid();
 
-  ora().info(`Generating the Typescript package "${core.consts.projectName}" at "${core.consts.projectPath}"...`);
+  ora().info(`Generating the Typescript project '${core.consts.projectName}' at '${core.consts.projectPath}'`);
 
   await core.actions.setProjectDirectory();
 
-  await core.actions.applyTemplate();
+  await core.actions.applySemitemplate();
 
   core.add.changelog();
   core.add.readme({
@@ -32,17 +51,10 @@ export const flavorTypescript: FlavorFunction = async (core) => {
 
   // To install the latest. The semitemplate deps don't matter too much,
   await core.actions.addPackages({
-    devDeps: [
-      'typescript@latest',
-      'ts-node-dev@latest',
-      'eslint@latest',
-      'eslint-config-gev@latest',
-      '@typescript-eslint/eslint-plugin@latest',
-      '@typescript-eslint/parser@latest',
-      '@types/node',
-      // 'ts-node@latest', // Add it for `once` script or wait https://github.com/wclr/ts-node-dev/issues/263
-    ],
+    devDeps: typescriptDevDeps,
   });
 
-  ora().succeed(`Package "${core.consts.projectName}" created at "${core.consts.projectPath}"!`);
+  ora().succeed(`Typescript project '${core.consts.projectName}' created at '${core.consts.projectPath}'!`);
 };
+
+export default flavorTypescript;