+++ src/typesAndConsts.ts
@@ -1,18 +1,12 @@
-import { Core } from './core';
+import { Core } from './core/core';
 import Path from 'path';
 
 
-export const flavorsArray = ['ts', 'expo', 'expo-pkg'] as const;
-
-
-export type Flavor = (typeof flavorsArray)[number]
-
-export function getFlavorWritePath(flavor: Flavor): string {
+export function getFlavorWritePath(flavor: string): string {
   return Path.join(__dirname, '..', 'semitemplates', flavor);
 }
 
 /** Accepts '.' or './' to inform to use the cwd */
 export const currentDirectoryRegex = /\.\.?/;
 
-
-export type FlavorFunction = (common: Core) => (void | Promise<void>)
+export type FlavorFunction = (common: Core) => (void | Promise<void>);